==> 0_my_key.R <==
# 0_my_key.R
# my API key; kept seperate as not for sharing
# July 2022

yourdatacenterid = 'syd1' #
base_url = 'syd1.qualtrics.com' # see https://api.qualtrics.com/ZG9jOjg3NjYzMw-base-url-and-datacenter-i-ds
key = 'Gp7xNY01SzOyRySv7kDOdUuz1cj6IgpffukTnPEV' # from My Account -> Qualtrics IDs
